Add proper series metadata tags
This takes the feed title if it is set and falls back to the ID. The ID
fallback might be a bit crude but I'd have to shuffle around a lot to
make the URL or feed name in the toml work.

Ideally I want the series to just be set from the TOML key.