Move the whole feed reader logic into more rust like types
This is a first iteration, I have had more thoughts since I wrote this
but it compiles and the core logic did not change so I am somewhat
confident in it working out fine. There's going to be a much bigger
rewrite of some of this to make it more testable now that I've proven
the concept and gathered some experience in what I want to do, so this
commit will probably be inconsequential.