Rewrite all of the core feed_reader module to be more testable
The main catch is that we've prepared for time travel in the main code
function since we now pass in an external timestamp that lets us control
what the code thinks is the current time, with that we can get eventual
testing for the whole feed_reader down to be just in unit tests.

This does still feel a bit odd around some types like `ConditionalType`
they are maybe in the wrong module I'm not sure the `config` module
really is the correct place. But the rewrite has gotten big enough where
I just wanna finish a part that works so it stops being broken at every
corner.